fix: move to main logging
