+++ persistence-modules/spring-boot-persistence-mongodb/src/main/java/com/baeldung/logging/model/Book.java
@@ -1,4 +1,4 @@
-package com.baeldung.mongodb.models;
+package com.baeldung.logging.model;
 
 import java.util.Objects;
